update FTerm config

+++ lua/plugins.lua
@@ -726,7 +726,7 @@ return require('packer').startup(function(use)
 		'numToStr/FTerm.nvim',
 		config = function ()
 			require'FTerm'.setup({
-				border = 'double',
+				border = 'single',
 				dimensions  = {
 					height = 0.9,
 					width = 0.9,
